Refactor AnalyzeLoop()
